fix: Move the logs-sdk and metrics-sdk requires
This makes it so the logs and metrics will work when you require like
```ruby
require 'opentelemetry/sdk/logs'
require 'opentelemetry/sdk/metrics'
```
Without getting 'opentelemetry-logs-api' or 'opentelemetry-logs-metrics'
earlier.

The opentelemetry-exporter-otlp-logs gem works around this, but I'm
leaving that workaround there for now, so it will work with old versions
of the opentelemetry-logs-sdk gem that are still locked.

[Fixes #1955]